fix: resolve 32-bit overflow issues (#6385)
- Use strconv.FormatInt with int64 cast for maxSafeInteger
- Explicitly cast uint32 limits to uint64 in error formatting

+++ internal/output/prometheusrw/remote/client.go
@@ -128,7 +128,7 @@ func newWriteRequestBody(series []*prompb.TimeSeries) ([]byte, error) {
 	}
 	if snappy.MaxEncodedLen(len(b)) < 0 {
 		return nil, fmt.Errorf("the protobuf message is too large to be handled by Snappy encoder; "+
-			"size: %d, limit: %d", len(b), math.MaxUint32)
+			"size: %d, limit: %d", len(b), uint64(math.MaxUint32))
 	}
 	return snappy.Encode(nil, b), nil
 }
